generated_code/2/agent2/script.py
import json
import sys
from appium import webdriver
from appium.webdriver.common.appiumby import AppiumBy

def run_automation(device_config_path):
    with open(device_config_path, 'r') as f:
        config = json.load(f)
    
    driver = webdriver.Remote('http://127.0.0.1:4723', config)
    driver.implicitly_wait(10)
    
    try:
        # Generated automation steps
        username_field = driver.find_element(AppiumBy.ID, "username")
        username_field.send_keys("test_user")
        
        password_field = driver.find_element(AppiumBy.ID, "password")
        password_field.send_keys("test_pass")
        
        login_button = driver.find_element(AppiumBy.ID, "login_button")
        login_button.click()
        
        print("Mobile automation completed successfully")
    except Exception as e:
        print(f"Mobile automation failed: {e}")
    finally:
        driver.quit()

if __name__ == "__main__":
    if len(sys.argv) != 2:
        print("Usage: python script.py <device_config.json>")
        sys.exit(1)
    run_automation(sys.argv[1])