generated_code/2/agent2/requirements.txt
Appium-Python-Client>=3.1.0
uiautomator2>=3.0.0